Move chat view tests to their own file
